use connId field in QueryCtx

+++ pkg/proxy/driver/domain.go
@@ -23,8 +23,8 @@ type Breaker interface {
 	GetBreakerScope() string
 	Hit(name string, idx int, isFail bool) error
 	Status(name string) (int32, int)
-	AddTimeWheelTask(name string, connectionID uint32, flag *int32) error
-	RemoveTimeWheelTask(connectionID uint32) error
+	AddTimeWheelTask(name string, connectionID uint64, flag *int32) error
+	RemoveTimeWheelTask(connectionID uint64) error
 	CASHalfOpenProbeSent(name string, idx int, halfOpenProbeSent bool) bool
 	CloseBreaker()
 }
